requirements.txt
pandas
scikit-learn
joblib

numpy
requests
flask
google.generativeai
flask-cors
